requirements.txt
django ~= 3.1.2
django-cors-headers ~= 3.5.0
django-extensions ~= 3.1.0
djangorestframework-jsonapi ~= 4.0.0
psycopg2-binary ~= 2.8.6

